+++ tests/matmul/program.bpf.c
@@ -6,7 +6,7 @@
 int main() {
     volatile float mat_1[MAT_SIZE], mat_2[MAT_SIZE], mat_res[MAT_SIZE];
     struct bpf_iter_num it;
-    int* i;
+    long long* i;
 
     bpf_iter_num_new(&it, 0, MAT_SIZE);
     while ((i = bpf_iter_num_next(&it))) {
@@ -18,9 +18,9 @@ int main() {
 
     bpf_iter_num_new(&it, 0, MAT_DIM * MAT_DIM * MAT_DIM);
     while ((i = bpf_iter_num_next(&it))) {
-        int row = *i / (MAT_DIM * MAT_DIM);
-        int col = (*i / MAT_DIM) % MAT_DIM;
-        int k = *i % MAT_DIM;
+        int row = (int)*i / (MAT_DIM * MAT_DIM);
+        int col = ((int)*i / MAT_DIM) % MAT_DIM;
+        int k = (int)*i % MAT_DIM;
         mat_res[row * MAT_DIM + col] += mat_1[row * MAT_DIM + k] * mat_2[k * MAT_DIM + col];
     }
     bpf_iter_num_destroy(&it);
